refactor: modularize and relocate entity and block interaction handlers
- Split `entityDie` and `entityInteract` into dedicated modules in `features/interactions`.
- Streamlined block interaction handling, reducing duplication.
- Updated references in `events/index.ts` to reflect new file locations.
