+++ behaviour_pack/scripts-dev/events/index.ts
@@ -1,8 +1,8 @@
-import load_block_event_handler from "./blocks";
+import load_block_event_handler from "../features/interactions/block-interact";
 import load_chat_handler from "./chat";
 import load_connections_handler from "./connections";
-import load_entity_event_handler from "./entities";
-import load_script_event_handler from "./script_events";
+import load_entity_event_handler from "../features/interactions/entity-die";
+import load_script_event_handler from "../features/interactions/script_events";
 import load_eliana_handler from "./eliana_trade";
 
 export default function load_world_event_handlers(guild_id: string) {
